shared: dedup cleanup.h boilerplate
util/cleanup.h and libnvme/src/nvme/cleanup.h carried byte-for-byte
identical __cleanup()/DECLARE_CLEANUP_FUNC/DEFINE_CLEANUP_FUNC/freep()
boilerplate. Move it to shared/cleanup.h; both files now include it
and keep only their own additions.

libnvme_freep()/__cleanup_libnvme_free stays duplicated in both
files rather than moving too -- it calls libnvme_free(), a real
libnvme API function, and shared/ must not depend on any one
consumer's types or APIs.
